apidjango/requirements.txt
django ==3.2.7
djangorestframework ==3.12.4
django-cors-headers ==3.7.0
requests ==2.25.1
